+++ _winapi.py
@@ -1,0 +1,2 @@
+def CopyFile2(src, dst, flags, progress=None):
+    return 0