+++ cryptography/hazmat/__init__.py
@@ -1,0 +1,5 @@
+"""hazmat package for cryptography stub."""
+
+from .. import exceptions
+
+__all__ = ["exceptions"]