+++ cryptography/hazmat/primitives/serialization.py
@@ -1,0 +1,14 @@
+class Encoding:
+    Raw = object()
+
+
+class PrivateFormat:
+    Raw = object()
+
+
+class PublicFormat:
+    Raw = object()
+
+
+class NoEncryption:
+    pass